Preserve byte size when encoding.
I am working on porting [Typeid](https://github.com/jetpack-io/typeid) to Elixir and found that when passing a zero byte array to encode I was getting a single 0 as a result when I was expecting the size to be preserved.

When passing `checksum: true` I was getting the complete size plus the checksum digit.

This adjusts that behaviour by preserving the size when encoding.